Move AnimationData from Component/ to Config/, close #156
