+++ Ash2/src/Config/AnimationData.hpp
